Handle empty file names (#10)
This commit enables the option to support queries on an empty path. The behavior expects that an empty file should always return and not handle the path. When an empty path is queried to know if it is firable, the engine will always return true, expecting the method to fail because the underlying implementation of the glibc actually tries to complete the operation.

Co-authored-by: Marco Edoardo Santimaria <marcoedoardo.santimaria@unito.it>